cleanByPeriodOffset add date compare

+++ logback-core/src/main/java/ch/qos/logback/core/rolling/helper/TimeBasedArchiveRemover.java
@@ -26,12 +26,19 @@ public class TimeBasedArchiveRemover extends DefaultArchiveRemover {
   protected void cleanByPeriodOffset(Date now, int periodOffset) {
     Date date2delete = rc.getRelativeDate(now, periodOffset);
     String filename = fileNamePattern.convert(date2delete);
+    
     File file2Delete = new File(filename);
+    
     if (file2Delete.exists() && file2Delete.isFile()) {
-      file2Delete.delete();
-      addInfo("deleting " + file2Delete);
-      if (parentClean) {
-        removeFolderIfEmpty(file2Delete.getParentFile());
+      Date fileLastModified = rc.getRelativeDate(new Date(file2Delete.lastModified()), -1);
+      
+      if (fileLastModified.compareTo(date2delete) <= 0) {
+        file2Delete.delete();
+        addInfo("deleting " + file2Delete);
+          
+        if (parentClean) {
+          removeFolderIfEmpty(file2Delete.getParentFile());
+        }
       }
     }
   }
